ast_results/xbmc_Kore/app/src/test/java/org/xbmc/kore/tests/jsonrpc/notifications/ApplicationTest.java
// isComment
package org.xbmc.kore.tests.jsonrpc.notifications;

import android.os.Handler;
import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.robolectric.RobolectricTestRunner;
import org.robolectric.annotation.Config;
import org.robolectric.shadows.ShadowLog;
import org.xbmc.kore.BuildConfig;
import org.xbmc.kore.host.HostInfo;
import org.xbmc.kore.jsonrpc.ApiCallback;
import org.xbmc.kore.jsonrpc.HostConnection;
import org.xbmc.kore.jsonrpc.notification.Application;
import org.xbmc.kore.testutils.tcpserver.MockTcpServer;
import org.xbmc.kore.testutils.tcpserver.handlers.ApplicationHandler;
import org.xbmc.kore.testutils.tcpserver.handlers.JSONConnectionHandlerManager;
import org.xbmc.kore.utils.RoboThreadRunner;
import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;
import static org.junit.Assert.fail;

@RunWith(RobolectricTestRunner.class)
@Config(constants = BuildConfig.class, sdk = isIntegerConstant)
public class isClassOrIsInterface {

    private HostConnection isVariable;

    private MockTcpServer isVariable;

    private ApplicationHandler isVariable;

    @Before
    public void isMethod() throws Exception {
        isNameExpr.isFieldAccessExpr = isNameExpr.isFieldAccessExpr;
        isNameExpr = new ApplicationHandler();
        JSONConnectionHandlerManager isVariable = new JSONConnectionHandlerManager();
        isNameExpr.isMethod(isNameExpr);
        isNameExpr = new MockTcpServer(isNameExpr);
        isNameExpr.isMethod();
        HostInfo isVariable = new HostInfo("isStringConstant", isNameExpr.isMethod(), isNameExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr, isNameExpr.isMethod(), null, null, true, isNameExpr.isFieldAccessExpr, true);
        isNameExpr = new HostConnection(isNameExpr);
    }

    @After
    public void isMethod() throws Exception {
        isNameExpr.isMethod();
        isNameExpr.isMethod();
    }

    @Test
    public void isMethod() throws InterruptedException {
        HostConnection.ApplicationNotificationsObserver isVariable = new HostConnection.ApplicationNotificationsObserver() {

            @Override
            public void isMethod(Application.OnVolumeChanged isParameter) {
                isNameExpr.isMethod();
                isMethod(isNameExpr.isFieldAccessExpr == isIntegerConstant);
            }
        };
        isNameExpr.isMethod(isNameExpr, new Handler());
        isNameExpr.isMethod(isIntegerConstant, true);
        isMethod(isIntegerConstant);
        isMethod(isNameExpr.isMethod(isIntegerConstant));
    }

    @Test
    public void isMethod() throws InterruptedException {
        HostConnection.ApplicationNotificationsObserver isVariable = new HostConnection.ApplicationNotificationsObserver() {

            @Override
            public void isMethod(Application.OnVolumeChanged isParameter) {
                isNameExpr.isMethod();
                isMethod(isNameExpr.isFieldAccessExpr);
            }
        };
        isNameExpr.isMethod(isNameExpr, new Handler());
        isNameExpr.isMethod(true, true);
        isMethod();
        isMethod(isNameExpr.isMethod(isIntegerConstant));
    }

    @Test
    public void isMethod() throws InterruptedException {
        HostConnection.ApplicationNotificationsObserver isVariable = new HostConnection.ApplicationNotificationsObserver() {

            @Override
            public void isMethod(Application.OnVolumeChanged isParameter) {
                isNameExpr.isMethod();
                isMethod(isNameExpr.isFieldAccessExpr);
            }
        };
        isNameExpr.isMethod(isNameExpr, new Handler());
        isNameExpr.isMethod(true, true);
        isMethod();
        isMethod(isNameExpr.isMethod(isIntegerConstant));
    }

    private void isMethod() {
        org.xbmc.kore.jsonrpc.method.Application.SetMute isVariable = new org.xbmc.kore.jsonrpc.method.Application.SetMute();
        isNameExpr.isMethod(isNameExpr, new ApiCallback<Boolean>() {

            @Override
            public void isMethod(Boolean isParameter) {
            }

            @Override
            public void isMethod(int isParameter, String isParameter) {
                isNameExpr.isMethod();
                isMethod("isStringConstant" + isNameExpr + "isStringConstant" + isNameExpr);
            }
        }, new Handler());
    }

    private void isMethod(int isParameter) {
        org.xbmc.kore.jsonrpc.method.Application.SetVolume isVariable = new org.xbmc.kore.jsonrpc.method.Application.SetVolume(isNameExpr);
        isNameExpr.isMethod(isNameExpr, new ApiCallback<Integer>() {

            @Override
            public void isMethod(Integer isParameter) {
            }

            @Override
            public void isMethod(int isParameter, String isParameter) {
                isNameExpr.isMethod();
                isMethod("isStringConstant" + isNameExpr + "isStringConstant" + isNameExpr);
            }
        }, new Handler());
    }
}
